pyproject.toml
[tool.poetry]
name = "crowdify-gw"
version = "1.0.0"
description = "Backend API para plataforma de venta de tickets para eventos"
authors = ["Crowdify Team"]
readme = "README.md"
packages = [{include = "app"}, {include = "services"}, {include = "shared"}, {include = "scripts"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
email-validator = "^2.2.0"
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.1"
redis = "^5.0.1"
celery = "^5.3.4"
flower = "^2.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
mercadopago = "^2.2.0"
sendgrid = "^6.11.0"
slowapi = "^0.1.9"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# Scripts útiles
generate-token = "scripts.generate_token:main"
